+++ CHANGELOG.md
@@ -111,6 +111,7 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
 ### Fixed
 
 ### Security
+- Secure cookies should contain the HTTPOnly attribute
 
 See the [full commit history](https://github.com/DFE-Digital/complete-conversions-transfers-changes/compare/production-2025-10-17.921...production-2025-10-20.943) for everything in the release